app/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pymysql==1.1.0
pydantic==2.6.1
python-dotenv==1.0.1
openai==1.12.0
cryptography==41.0.3
alembic==1.18.4
groq==1.0.0